Fixed a bug where npc healthbar health value would disappear.
This bug would only be encountered when the healthbar plate moved past
the 16:9 boundary on the left side on ultrawide displays. 32:9 would
see it more profoundly.

When health bars moved past the boundary the health value of npc's
would just disappear, this was caused because the variable that stored
the position of the healtbar was a u64 when it should have been f32, as
the value we are saving is coming from a xmm fpu register. When the
variable was a u64 and the health bar went past the 16:9 boundary on
the left it would become a very large number because past the 16:9
boundary on the left the numbers would be negative and because it was
declared as u64 that negative number was interpreted as a very big
positive number, just by changing it to i64 it would have fixed the
issue, but ultimately as stated before f32 is most correct as we are
taking this value from an fpu register.

+++ src/dllmain.cpp
@@ -293,7 +293,7 @@ void fixResolution() {
 void fixHud() {
     static bool isHealthBarTrack = false;
     static bool isHealthBarFill = false;
-    static u64 healthBarTrackX = 0;
+    static f32 healthBarTrackX = 0;
     static bool glpyhStringEnd = false;
     static std::vector<u64> glyphString{};
     Utils::SignatureHook hook1 {
